Further refactoring, Sherlock takes shape.
