+++ Bricks/template/is_unique_ptr.h
@@ -56,6 +56,42 @@ static_assert(std::is_same<int, typename is_unique_ptr<int>::underlying_type>::v
 static_assert(is_unique_ptr<std::unique_ptr<int>>::value, "");
 static_assert(std::is_same<int, typename is_unique_ptr<std::unique_ptr<int>>::underlying_type>::value, "");
 
+
+template <typename B, typename E>
+struct can_be_stored_in_unique_ptr {
+  static constexpr bool value = false;
+};
+
+// Note: No custom deleters here [yet]. TODO(dkorolev): Make sure they are here as we need them.
+template <typename B, typename E>
+struct can_be_stored_in_unique_ptr<std::unique_ptr<B>, E> {
+  static constexpr bool value = std::is_same<B, E>::value || std::is_base_of<B, E>::value;
+};
+
+namespace selftest {
+
+struct BASE {
+  virtual ~BASE() = default;
+};
+
+struct DERIVED : BASE {};
+
+struct STANDALONE {};
+
+static_assert(can_be_stored_in_unique_ptr<std::unique_ptr<BASE>, BASE>::value, "");
+static_assert(can_be_stored_in_unique_ptr<std::unique_ptr<BASE>, DERIVED>::value, "");
+static_assert(!can_be_stored_in_unique_ptr<std::unique_ptr<BASE>, STANDALONE>::value, "");
+
+static_assert(!can_be_stored_in_unique_ptr<std::unique_ptr<DERIVED>, BASE>::value, "");
+static_assert(can_be_stored_in_unique_ptr<std::unique_ptr<DERIVED>, DERIVED>::value, "");
+static_assert(!can_be_stored_in_unique_ptr<std::unique_ptr<BASE>, STANDALONE>::value, "");
+
+static_assert(!can_be_stored_in_unique_ptr<std::unique_ptr<BASE>, std::unique_ptr<BASE>>::value, "");
+static_assert(!can_be_stored_in_unique_ptr<std::unique_ptr<BASE>, std::unique_ptr<DERIVED>>::value, "");
+static_assert(!can_be_stored_in_unique_ptr<std::unique_ptr<DERIVED>, std::unique_ptr<DERIVED>>::value, "");
+
+}  // namespace bricks::selftest
+
 }  // namespace bricks
 
 #endif  // BRICKS_TEMPLATE_IS_UNIQUE_PTR_H
